moving things around

+++ src/main/java/net/tackley/weretwit/moon/Moon.java
@@ -3,5 +3,5 @@ package net.tackley.weretwit.moon;
 import java.util.Date;
 
 public interface Moon {
-    public int getPhaseOfMoon(Date date);
+    public double getPhaseOfMoon(Date date);
 }
